process-application/src/WarehouseRobotPythonScripts/requirements.txt
basic-auth==0.0.2
boto==2.49.0
camunda-external-task-client-python3==4.3.0
certifi==2022.12.7
chardet==3.0.4
DateTime==5.0
dnspython==2.3.0
ez-setup==0.9
ftrobopy==1.80
idna==2.10
mypy==0.991
mypy-extensions==0.4.3
mysql==0.0.3
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.1
pandas==1.5.3
protobuf==3.20.3
pydantic==1.10.4
pygame==2.1.3.dev8
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.24.0
six==1.16.0
tomli==2.0.1
translationstring==1.4
typing_extensions==4.4.0
urllib3==1.25.11
zope.interface==5.5.2